Add Localization by operator diag method

+++ src/surface_potential_analysis/surface_potential_analysis/potential/potential.py
@@ -300,7 +300,7 @@ def interpolate_uneven_potential(
     util = BasisUtil((data["basis"][0], data["basis"][1]))
     # TODO: maybe along axis
     xy_interpolated = interpolate_points_rfftn(
-        data["vector"].reshape(*util.shape, len(data["basis"][2])),
+        data["vector"].reshape(*util.shape, len(data["basis"][2])).astype(np.float_),
         s=(shape[0], shape[1]),
         axes=(0, 1),
     )
